configure: swap lib order in LIBMXF_LDADDLIBS
The wrapper script used for executables and generated by libtool
sets the LD_LIBRARY_PATH before running the executable. Prior to
this commit it would set the local library path (src/.libs) after
the /usr/local/lib when present.

This was not an issue in libMXF but it brings it inline with bmx where
it was.